secondDay
test fix + OOP so hard
