+++ src/test/java/pageObjects/PageMethods.java
@@ -1,11 +1,14 @@
-package actions;
+package pageObjects;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.ui.WebDriverWait;
 
-import pageObjects.PageObjects;
+import utility.Root;
 
-public class WebElements extends PageObjects {
+public class PageMethods extends Root {
+
+	protected WebDriverWait wait = new WebDriverWait(driver, 10);
 
 	public WebElement let(By locator) {
 		WebElement Element = driver.findElement(locator);
@@ -15,4 +18,5 @@ public class WebElements extends PageObjects {
 	public void log(String element) {
 		System.out.println(element);
 	}
+
 }
